+++ asset_hashes_gen.go
@@ -181,6 +181,6 @@ var precomputedAssetHashes = map[string]string{
 	"js/visual-settings.js":                                  "5e05fd4f5ab7",
 	"js/weather.js":                                          "2e4cddbdcd40",
 	"js/whats-new-modal.js":                                  "e7fdc7a5c8d5",
-	"js/whats-new-stub.js":                                   "7cfe4a52dab3",
+	"js/whats-new-stub.js":                                   "e9f64173b799",
 	"js/write-api.js":                                        "25afcd408905",
 }